mesa: patch application overrides in drirc for asahi devices
Adds a programtic xmlstarlet snippet to record all the names of the
application overrides for asahi devices and duplicate them for the
wrapped executables. It's required for nix to have the "wrapped"
executable names also present in the mesa overrides drirc file for the
override to go into effect.

The reason to use a programatic xmlstarlet script is to avoid having
to track a patch file that may break on future releases. This snippet
should always work no matter the formatting of the
`00-mesa-defaults.conf` file.